+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Preview app in VS Code Editor added as an option
 - Option to install new packages
 - Option to uninstall packages
+- Allow JAVA_HOME to be set (to allow building with different version of Java)
 - Settings are now applied immediated (do not require a refresh)
 - Build command now uses ionic build
 - Production flag setting now correctly works